requirements.txt
# OncoAgent Dependencies
# =====================

# LangGraph - Orchestration framework
langgraph>=0.2.0

# LangChain Core
langchain>=0.3.0
langchain-core>=0.3.0

# LangChain Integrations
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0

# Exa Search
langchain-exa>=0.2.0
exa-py>=1.0.0

# Pydantic for structured outputs
pydantic>=2.0.0

# Optional: For visualization
# pygraphviz  # Requires system install of graphviz

# Optional: For tracing/debugging
# langsmith>=0.1.0
# Utilities
python-dotenv>=1.0.0

# FastAPI
fastapi>=0.110.0
uvicorn>=0.27.0

# Streamlit UI
streamlit>=1.31.0

# Database
psycopg2-binary>=2.9.9
langgraph-checkpoint-postgres>=0.1.0

# HTTP
requests>=2.31.0
httpx>=0.27.0

# Google AI
google-genai>=1.0.0
